TalonSlavingFix + Auton
TalonSlavingFix - Talon 3 or rearLeft motor was not set as a follower to the Left front motor and was not recieving a value causing the the left side to run slower than the right side.  Auton - Development for the Auton to simulate for the 2018 Frc robot in development using a PID as a way to move forward. The Code has not been tested and does not have presices measurment or numbers inputted into the code.