TalonFix + Auton
Talon 3 has been fixed and is now set to follower mode and is slaved to the leftFrontBase. The Auton has made progress. The Drive forward command need serious revising due to it not actual moving forward just jerking back and forth. A turn command has been added to the Auton and is working.